+++ Sources/App/Helper/Writer/JsonWriter.swift
@@ -1,0 +1,186 @@
+import Foundation
+import Vapor
+
+extension BodyStreamWriter {
+    /// Execute async code and capture any errors. In case of error, print the error to the output stream
+    func submit(_ task: @escaping () async throws -> Void) {
+        _ = eventLoop.performWithTask { try await task() }
+            .flatMapError({ error in
+                write(.buffer(.init(string: "Unexpected error while streaming data: \(error)")))
+                    .flatMap({write(.error(error))})
+        })
+    }
+}
+
+extension Array where Element == () throws -> ForecastapiResult {
+    /**
+     Stream a potentially very large resultset to the client. The JSON file could easily be 20 MB.
+     Instead of generating a massive string in memory, we only allocate 18kb and flush every time the buffer exceeds 16kb.
+     Memory footprint is therefore much smaller and fits better into L2/L3 caches.
+     Additionally code is fully async, to not block the a thread for almost a second to generate a JSON response...
+     */
+    func toJsonResponse() throws -> Response {
+        // First excution outside stream, to capture potential errors better
+        var first = try self.first?()
+        
+        let response = Response(body: .init(stream: { writer in
+            writer.submit {
+                var b = BufferAndWriter(writer: writer)
+                /// For multiple locations, create an array of results
+                let isMultiPoint = count > 1
+                if isMultiPoint {
+                    b.buffer.writeString("[")
+                }
+                if let first {
+                    try await first.streamJsonResponse(to: &b)
+                }
+                first = nil
+                for closure in self.dropFirst() {
+                    let result = try closure()
+                    b.buffer.writeString(",")
+                    try await result.streamJsonResponse(to: &b)
+                }
+                if isMultiPoint {
+                    b.buffer.writeString("]")
+                }
+                try await b.flush()
+                try await b.end()
+            }
+        }))
+        response.headers.replaceOrAdd(name: .contentType, value: "application/json; charset=utf-8")
+        return response
+    }
+}
+
+extension ForecastapiResult {
+    /**
+     Stream a potentially very large resultset to the client. The JSON file could easily be 20 MB.
+     Instead of generating a massive string in memory, we only allocate 18kb and flush every time the buffer exceeds 16kb.
+     Memory footprint is therefore much smaller and fits better into L2/L3 caches.
+     Additionally code is fully async, to not block the a thread for almost a second to generate a JSON response...
+     */
+     func toJsonResponse() -> Response {
+        let response = Response(body: .init(stream: { writer in
+            _ = writer.eventLoop.performWithTask {
+                var b = BufferAndWriter(writer: writer)
+                try await streamJsonResponse(to: &b)
+                try await b.end()
+            }
+            
+        }, count: -1))
+        
+        response.headers.replaceOrAdd(name: .contentType, value: "application/json; charset=utf-8")
+        return response
+    }
+    
+    fileprivate func streamJsonResponse(to b: inout BufferAndWriter) async throws {
+        b.buffer.writeString("""
+        {"latitude":\(latitude),"longitude":\(longitude),"generationtime_ms":\(generationtime_ms),"utc_offset_seconds":\(utc_offset_seconds),"timezone":"\(timezone.identifier)","timezone_abbreviation":"\(timezone.abbreviation() ?? "")"
+        """)
+        if let elevation = elevation, elevation.isFinite {
+            b.buffer.writeString(",\"elevation\":\(elevation)")
+        }
+        if let current_weather = current_weather {
+            let ww = current_weather.weathercode.isFinite ? String(format: "%.0f", current_weather.weathercode) : "null"
+            let is_day = current_weather.is_day.isFinite ? String(format: "%.0f", current_weather.is_day) : "null"
+            let winddirection = current_weather.winddirection.isFinite ? String(format: "%.0f", current_weather.winddirection) : "null"
+            b.buffer.writeString("""
+                ,"current_weather":{"temperature":\(current_weather.temperature),"windspeed":\(current_weather.windspeed),"winddirection":\(winddirection),"weathercode":\(ww),"is_day":\(is_day),"time":
+                """)
+            b.buffer.writeString(current_weather.time.formated(format: timeformat, utc_offset_seconds: utc_offset_seconds, quotedString: true))
+            b.buffer.writeString("}")
+        }
+        
+        /// process sections like hourly or daily
+        for section in sections {
+            b.buffer.writeString(",\"\(section.name)_units\":")
+            b.buffer.writeString("{")
+            switch timeformat {
+            case .iso8601:
+                b.buffer.writeString("\"time\":\"\(SiUnit.iso8601.rawValue)\"")
+            case .unixtime:
+                b.buffer.writeString("\"time\":\"\(SiUnit.unixtime.rawValue)\"")
+            }
+            for e in section.columns {
+                b.buffer.writeString(",\"\(e.variable)\":\"\(e.unit.rawValue)\"")
+                try await b.flushIfRequired()
+            }
+            b.buffer.writeString("}")
+            b.buffer.writeString(",\"\(section.name)\":")
+            b.buffer.writeString("{")
+            b.buffer.writeString("\"time\":[")
+            
+            // Write time axis
+            var firstValue = true
+            for time in section.time.itterate(format: timeformat, utc_offset_seconds: utc_offset_seconds, quotedString: true, onlyDate: section.time.dtSeconds == 86400) {
+                if firstValue {
+                    firstValue = false
+                } else {
+                    b.buffer.writeString(",")
+                }
+                b.buffer.writeString(time)
+                try await b.flushIfRequired()
+            }
+            b.buffer.writeString("]")
+            
+            /// Write data
+            for e in section.columns {
+                b.buffer.writeString(",")
+                b.buffer.writeString("\"\(e.variable)\":")
+                b.buffer.writeString("[")
+                var firstValue = true
+                switch e.data {
+                    /*case .string(let strings):
+                     for v in strings {
+                     if firstValue {
+                     firstValue = false
+                     } else {
+                     b.buffer.writeString(",")
+                     }
+                     b.buffer.writeString("\"\(v)\"")
+                     try await b.flushIfRequired()
+                     }*/
+                case .float(let floats):
+                    let format = "%.\(e.unit.significantDigits)f"
+                    for v in floats {
+                        if firstValue {
+                            firstValue = false
+                        } else {
+                            b.buffer.writeString(",")
+                        }
+                        if v.isFinite {
+                            b.buffer.writeString(String(format: format, v))
+                        } else {
+                            b.buffer.writeString("null")
+                        }
+                        try await b.flushIfRequired()
+                    }
+                    /*case .int(let ints):
+                     for v in ints {
+                     if firstValue {
+                     firstValue = false
+                     } else {
+                     b.buffer.writeString(",")
+                     }
+                     b.buffer.writeString("\(v)")
+                     try await b.flushIfRequired()
+                     }*/
+                case .timestamp(let timestamps):
+                    for time in timestamps.itterate(format: timeformat, utc_offset_seconds: utc_offset_seconds, quotedString: true, onlyDate: false) {
+                        if firstValue {
+                            firstValue = false
+                        } else {
+                            b.buffer.writeString(",")
+                        }
+                        b.buffer.writeString(time)
+                        try await b.flushIfRequired()
+                    }
+                }
+                b.buffer.writeString("]")
+                try await b.flushIfRequired()
+            }
+            b.buffer.writeString("}")
+        }
+        b.buffer.writeString("}")
+    }
+}
